Securisation CommentController: - modification de la route et maj template - création d'une méthode pour obtenir un episode à partir des paramètres de la route - new ROLE_SUBSCRIBER - edit et delete pour l'author ou ROLE_ADMIN - test associé

+++ src/Controller/CommentController.php
@@ -3,10 +3,10 @@
 namespace App\Controller;
 
 use App\Entity\Comment;
-use App\Entity\Episode;
 use App\Entity\User;
 use App\Form\CommentType;
-use App\Repository\CommentRepository;
+use App\Repository\EpisodeRepository;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -14,24 +14,15 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
- * @Route("episode/{id}/comment")
+ * @Route("program/{programSlug}/season/{seasonNumber}/episode/{episodeNumber}/comment")
  */
 class CommentController extends AbstractController
 {
     /**
-     * @Route("/", name="comment_index", methods={"GET"})
-     */
-    public function index(CommentRepository $commentRepository): Response
-    {
-        return $this->render('comment/index.html.twig', [
-            'comments' => $commentRepository->findAll(),
-        ]);
-    }
-
-    /**
      * @Route("/new", name="comment_new", methods={"GET","POST"})
+     * @IsGranted("ROLE_SUBSCRIBER")
      */
-    public function new(Request $request, Episode $episode): Response
+    public function new(Request $request, EpisodeRepository $episodeRepository): Response
     {
         /** @var User $user */
         $user = $this->getUser();
@@ -39,6 +30,8 @@ class CommentController extends AbstractController
         if (!($user instanceof User)) {
             return $this->redirectToRoute('app_login');
         }
+
+        $episode = $episodeRepository->findEpisodeByUrlParameter($request->attributes);
 
         $comment = new Comment($episode, $user);
         $form = $this->createForm(CommentType::class, $comment);
@@ -59,22 +52,16 @@ class CommentController extends AbstractController
     }
 
     /**
-     * @Route("/{comment_id}", name="comment_show", methods={"GET"})
-     * @ParamConverter("comment", options={"id" = "comment_id"})
-     */
-    public function show(Comment $comment): Response
-    {
-        return $this->render('comment/show.html.twig', [
-            'comment' => $comment,
-        ]);
-    }
-
-    /**
      * @Route("/{comment_id}/edit", name="comment_edit", methods={"GET","POST"})
      * @ParamConverter("comment", options={"id" = "comment_id"})
+     * @IsGranted("ROLE_SUBSCRIBER")
      */
     public function edit(Request $request, Comment $comment): Response
     {
+        if (!$this->isAuthorOrAdmin($comment)) {
+            return $this->redirectToRoute('app_login');
+        }
+
         $form = $this->createForm(CommentType::class, $comment);
         $form->handleRequest($request);
 
@@ -93,15 +80,31 @@ class CommentController extends AbstractController
     /**
      * @Route("/{comment_id}", name="comment_delete", methods={"DELETE"})
      * @ParamConverter("comment", options={"id" = "comment_id"})
+     * @IsGranted("ROLE_SUBSCRIBER")
      */
     public function delete(Request $request, Comment $comment): Response
     {
+        if (!$this->isAuthorOrAdmin($comment)) {
+            return $this->redirectToRoute('app_login');
+        }
+
         if ($this->isCsrfTokenValid('delete' . $comment->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($comment);
             $entityManager->flush();
         }
 
-        return $this->redirectToRoute('episode_show', ['id' => $comment->getEpisode()->getId()]);
+        return $this->redirectToRoute('episode_show', [
+            'programSlug' => $comment->getEpisode()->getSeason()->getProgram()->getSlug(),
+            'seasonNumber'=> $comment->getEpisode()->getSeason()->getNumber(),
+            'number'      => $comment->getEpisode()->getNumber(), ]);
+    }
+
+    private function isAuthorOrAdmin(Comment $comment): bool
+    {
+        /** @var User $user */
+        $user = $this->getUser();
+
+        return $this->isGranted('ROLE_ADMIN') || $comment->getAuthor() === $user;
     }
 }
